moved cookie consents in utils and called from script.js
